Apply community 26.1.2 port reference diff (base for 26.2 port)

+++ common/src/main/kotlin/com/cobblemon/mod/common/api/riding/behaviour/types/air/JetBehaviour.kt
@@ -443,7 +443,7 @@ class JetBehaviour : RidingBehaviour<JetSettings, JetState> {
         impactVec: Vec3
     ): Boolean {
         val impactSpeed = impactVec.horizontalDistance().toFloat() * 10f
-        return vehicle.causeFallDamage(impactSpeed, 1f, vehicle.damageSources().flyIntoWall())
+        return vehicle.causeFallDamage(impactSpeed.toDouble(), 1f, vehicle.damageSources().flyIntoWall())
     }
 
     override fun asMoLangValue(
