+++ common/src/main/kotlin/com/cobblemon/mod/common/client/render/models/blockbench/animation/PunchAnimation.kt
@@ -51,15 +51,16 @@ class PunchAnimation(
     ): Boolean {
         val attackTime = state.animationSeconds - startedTime
         var g: Float = attackTime
-        this.body.yRot = Mth.sin(Mth.sqrt(g) * (Math.PI * 2).toFloat()) * 0.2f
+        // PT128: Mth.sin/cos in MC 26.1.x take Double (was Float).
+        this.body.yRot = Mth.sin((Mth.sqrt(g) * (Math.PI * 2).toFloat()).toDouble()) * 0.2f
         if (!swingRight) {
             this.body.yRot *= -1.0f
         }
 
-        this.rightArm.z = Mth.sin(this.body.yRot) * 5.0f
-        this.rightArm.x = -Mth.cos(this.body.yRot) * 5.0f
-        this.leftArm.z = -Mth.sin(this.body.yRot) * 5.0f
-        this.leftArm.x = Mth.cos(this.body.yRot) * 5.0f
+        this.rightArm.z = Mth.sin((this.body.yRot).toDouble()) * 5.0f
+        this.rightArm.x = -Mth.cos((this.body.yRot).toDouble()) * 5.0f
+        this.leftArm.z = -Mth.sin((this.body.yRot).toDouble()) * 5.0f
+        this.leftArm.x = Mth.cos((this.body.yRot).toDouble()) * 5.0f
         this.rightArm.yRot += this.body.yRot
         this.leftArm.yRot += this.body.yRot
         this.leftArm.xRot += this.body.yRot
@@ -67,12 +68,12 @@ class PunchAnimation(
         g *= g
         g *= g
         g = 1.0f - g
-        val h = Mth.sin(g * Math.PI.toFloat())
-        val i: Float = Mth.sin(attackTime * Math.PI.toFloat()) * -(this.head.xRot - 0.7f) * 0.75f
+        val h = Mth.sin((g * Math.PI.toFloat()).toDouble())
+        val i: Float = Mth.sin((attackTime * Math.PI.toFloat()).toDouble()) * -(this.head.xRot - 0.7f) * 0.75f
         val modelPart = if (swingRight) this.rightArm else this.leftArm
         modelPart.xRot -= h * 1.2f + i
         modelPart.yRot += this.body.yRot * 2.0f
-        modelPart.zRot += Mth.sin(attackTime * Math.PI.toFloat()) * -0.4f
+        modelPart.zRot += Mth.sin((attackTime * Math.PI.toFloat()).toDouble()) * -0.4f
         return attackTime < duration
     }
 }
